Move labs serializer to shared space
